requirements.txt
boto3==1.35.6
botocore==1.35.6
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==1.26.19
